[STORE] move standby runtime and HA metrics under ha
