+++ mooncake-store/include/ha/metrics/ha_metric_manager.h
