+++ mooncake-store/include/ha/standby/standby_state_machine.h
