+++ mooncake-store/src/ha/standby/hot_standby_service.cpp
@@ -1,4 +1,4 @@
-#include "hot_standby_service.h"
+#include "ha/standby/hot_standby_service.h"
 
 #include <glog/logging.h>
 
@@ -6,7 +6,7 @@
 #include <chrono>
 #include <thread>
 
-#include "ha_metric_manager.h"
+#include "ha/metrics/ha_metric_manager.h"
 #include "ha/oplog/oplog_applier.h"
 #include "ha/oplog/oplog_codec.h"
 #include "ha/oplog/oplog_manager.h"
